// idfinv/src/app/shared/models/relascopica-semplice.ts
/*!
 * SPDX-FileCopyrightText: (C) Copyright 2023 Regione Piemonte
 * 
 * SPDX-License-Identifier: EUPL-1.2
 */
export class RelascopicaSemplice {
  specie?: string;
  note?: string;
  codiceAds?: string;
  comune?: string;
  alberiSeme?: number;
  alberiPollone?: number;
  tipologia?: string;
  fattoreNumerazione?: number;
  categoriaForestale?: string;
  dataRilevamento?: string;
  espozione?: string;
  inclinazione?: string;
  particella?: string;
  proprieta?: string;
  tipoForestale?: string;
  quota?: number;
  tipoStrutturale?: string;
  tipo?: string;
  diametro?: number;
  altezza?: number;
  incremento?: number;
  codTipoCampione?: string;
  utmEst?:number;
  utmNord?:number;
}